refactor: extract bootstrap wiring modules & table-drive undo snapshots
- Extract _wire_hierarchy_callbacks -> bootstrap_hierarchy.py (617 lines)
- Extract _wire_inspector_callbacks -> bootstrap_inspector.py (1032 lines)
- Extract _wire_project_callbacks -> bootstrap_project.py (220 lines)
- bootstrap.py reduced from 2931 -> 1100 lines (-62%)
- Table-drive undo.py snapshot/restore with _SNAPSHOT_REGISTRY
- Exclude test files from wheel (MANIFEST.in + pyproject.toml)
- Formalize Numba JIT kernels and stubs
- Add fps_stress_test, game_builder i18n, nuitka builder fixes
