+++ MANIFEST.in
@@ -1,1 +1,4 @@
 recursive-include python/Infernux *
+recursive-exclude python/Infernux/test *
+recursive-exclude python/test *
+global-exclude __pycache__ *.pyc *.pyo
