+++ pyproject.toml
@@ -40,7 +40,7 @@ Issues = "https://github.com/ChenlizheMe/Infernux/issues"
 launcher = ["PySide6>=6.5.0"]
 
 [tool.setuptools]
-packages = {find = {where = ["python"], exclude = ["*.__pycache__", "*.__pycache__.*"]}}
+packages = {find = {where = ["python"], exclude = ["test", "test.*", "*.test", "*.test.*", "*.tests", "*.tests.*", "*.__pycache__", "*.__pycache__.*"]}}
 package-dir = {"" = "python"}
 include-package-data = true
