+++ python/Infernux/engine/bootstrap_project.py
@@ -1,0 +1,220 @@
+"""
+Project panel callback wiring — extracted from EditorBootstrap.
+
+Provides :func:`wire_project_callbacks` which attaches all Python-side
+callbacks to a C++ ``ProjectPanel`` instance.
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from Infernux.engine.bootstrap import EditorBootstrap
+
+
+def wire_project_callbacks(bs: EditorBootstrap) -> None:
+    """Wire C++ ProjectPanel callbacks to Python managers."""
+    pp = bs.project_panel
+    from Infernux.engine.i18n import t as _t
+    from Infernux.engine.ui import project_file_ops as file_ops
+    from Infernux.engine.ui import project_utils
+    from Infernux.engine.scene_manager import SceneFileManager
+
+    # -- Engine subsystems --
+    native_engine = bs.engine.get_native_engine()
+    if native_engine:
+        pp.setup_from_engine(native_engine)
+
+    pp.set_root_path(bs.project_path)
+
+    import Infernux.resources as _resources
+    pp.set_icons_directory(_resources.file_type_icons_dir)
+
+    # -- Translation --
+    pp.translate = _t
+
+    # -- Asset database access (via engine) --
+    adb = bs.engine.get_asset_database()
+
+    pp.get_guid_from_path = lambda path: (
+        adb.get_guid_from_path(path) if adb else ""
+    )
+    pp.get_path_from_guid = lambda guid: (
+        adb.get_path_from_guid(guid) if adb else ""
+    )
+
+    # -- File operation callbacks --
+    from Infernux.debug import Debug
+
+    def _safe_project_create(cb, *args):
+        try:
+            return cb(*args)
+        except Exception as exc:
+            Debug.log_error(f"ProjectPanel create failed: {exc}")
+            return False, str(exc)
+
+    def _safe_project_path(cb, *args):
+        try:
+            return cb(*args) or ""
+        except Exception as exc:
+            Debug.log_error(f"ProjectPanel path operation failed: {exc}")
+            return ""
+
+    pp.create_folder = lambda cur, name: _safe_project_create(
+        file_ops.create_folder, cur, name)
+    pp.create_script = lambda cur, name: _safe_project_create(
+        file_ops.create_script, cur, name, adb)
+    pp.create_shader = lambda cur, name, typ: _safe_project_create(
+        file_ops.create_shader, cur, name, typ, adb)
+    pp.create_material = lambda cur, name: _safe_project_create(
+        file_ops.create_material, cur, name, adb)
+    pp.create_scene = lambda cur, name: _safe_project_create(
+        file_ops.create_scene, cur, name, adb)
+    pp.do_rename = lambda old, new_name: _safe_project_path(
+        file_ops.do_rename, old, new_name, adb)
+    pp.get_unique_name = lambda cur, base, ext: (
+        file_ops.get_unique_name(cur, base, ext)
+    )
+    pp.move_item_to_directory = lambda item, dest: _safe_project_path(
+        file_ops.move_item_to_directory, item, dest, adb)
+
+    # -- Delete (with Win32 confirmation dialog) --
+    def _delete_items(paths):
+        import ctypes, os
+        valid = []
+        seen = set()
+        for p in paths or []:
+            if not p or not os.path.exists(p) or p in seen:
+                continue
+            seen.add(p)
+            valid.append(p)
+        if not valid:
+            return
+
+        title = _t("project.delete_confirm_title")
+        if len(valid) == 1:
+            msg = _t("project.delete_confirm_msg").replace(
+                "{name}", os.path.basename(valid[0]))
+        else:
+            msg = _t("project.delete_confirm_multi_msg").replace(
+                "{count}", str(len(valid)))
+        # MB_OKCANCEL | MB_ICONWARNING | MB_DEFBUTTON2
+        result = ctypes.windll.user32.MessageBoxW(
+            0, msg, title, 0x1 | 0x30 | 0x100)
+        if result != 1:  # IDOK
+            return
+
+        for item_path in sorted(
+            valid, key=lambda p: (p.count(os.sep), len(p)), reverse=True
+        ):
+            if os.path.exists(item_path):
+                file_ops.delete_item(item_path, adb)
+
+    pp.delete_items = _delete_items
+
+    # -- Create prefab from hierarchy drag --
+    def _create_prefab_from_hierarchy(obj_id, current_path):
+        from Infernux.lib import SceneManager
+        scene = SceneManager.instance().get_active_scene()
+        if not scene:
+            return
+        game_object = scene.find_by_id(obj_id)
+        if not game_object:
+            return
+        ok, result = file_ops.create_prefab_from_gameobject(
+            game_object, current_path, adb)
+        if not ok:
+            from Infernux.debug import Debug
+            Debug.log_warning(f"Failed to create prefab: {result}")
+
+    pp.create_prefab_from_hierarchy = _create_prefab_from_hierarchy
+
+    # -- Open callbacks --
+    pp.open_file = lambda path: project_utils.open_file_with_system(
+        path, project_root=bs.project_path)
+
+    def _open_scene(file_path):
+        from Infernux.debug import Debug
+        from Infernux.engine.deferred_task import DeferredTaskRunner
+        from Infernux.engine.play_mode import PlayModeManager
+
+        def _open_after_stop():
+            sfm = SceneFileManager.instance()
+            if sfm:
+                return bool(sfm.open_scene(file_path))
+            Debug.log_warning("SceneFileManager not initialized")
+            return False
+
+        play_mode = PlayModeManager.instance()
+        if play_mode and play_mode.is_playing:
+            runner = DeferredTaskRunner.instance()
+            if runner.is_busy:
+                Debug.log_warning(
+                    "Cannot open scene while another deferred task is running")
+                return
+
+            def _on_stop(ok):
+                if not ok:
+                    Debug.log_warning(
+                        "Play Mode stop did not complete; scene open cancelled")
+                    return
+                try:
+                    from Infernux.lib import SceneManager as NativeSM
+                    nsm = NativeSM.instance()
+                except Exception:
+                    nsm = None
+                if play_mode.is_playing:
+                    Debug.log_warning(
+                        "Scene open cancelled — Play Mode still active")
+                    return
+                if nsm and nsm.is_playing():
+                    Debug.log_warning(
+                        "Scene open cancelled — native Play Mode still active")
+                    return
+                _open_after_stop()
+
+            if not play_mode.exit_play_mode(on_complete=_on_stop):
+                Debug.log_warning(
+                    "Failed to stop Play Mode before opening scene")
+            return
+
+        sfm = SceneFileManager.instance()
+        if sfm:
+            sfm.open_scene(file_path)
+        else:
+            Debug.log_warning("SceneFileManager not initialized")
+
+    pp.open_scene = _open_scene
+
+    pp.open_prefab_mode = lambda path: (
+        SceneFileManager.instance().open_prefab_mode_with_undo(path)
+        if SceneFileManager.instance() else None
+    )
+
+    pp.reveal_in_explorer = lambda path: (
+        project_utils.reveal_in_file_explorer(path)
+    )
+
+    # -- Script validation for drag-drop --
+    def _validate_script_component(file_path):
+        try:
+            from Infernux.components.script_loader import (
+                load_component_from_file, ScriptLoadError)
+            load_component_from_file(file_path)
+            return True
+        except Exception:
+            return False
+
+    pp.validate_script_component = _validate_script_component
+
+    # -- Inspector invalidation --
+    def _invalidate_asset_inspector(path):
+        try:
+            from Infernux.engine.ui.asset_inspector import (
+                invalidate_asset)
+            invalidate_asset(path)
+        except Exception:
+            pass
+
+    pp.invalidate_asset_inspector = _invalidate_asset_inspector
