+++ python/Infernux/jit.pyi
@@ -15,6 +15,13 @@ def njit(*args: Any, **kwargs: Any) -> Any:
 
     The decorated function gains a ``.py`` attribute pointing to the
     original pure-Python source.
+
+    Supports ``auto_parallel=True`` as an Infernux extension. In that mode
+    the wrapper prepares both serial and ``parallel=True`` variants,
+    conservatively upgrades simple ``for ... in range(...)`` reduction loops
+    to ``prange`` when possible, defaults to the parallel one, and lets
+    :func:`warmup` first trigger compilation and then benchmark steady-state
+    runtime before pinning the faster choice.
     """
     ...
